requirements.txt
Flask==2.3.3
Flask-Cors==6.0.2
python-dotenv==1.2.2
python-box==6.0.2
PyYAML==6.0.3
Pillow==11.3.0
numpy==1.23.5; python_version < "3.12"
numpy>=1.26,<2.2; python_version >= "3.12"
setuptools<81
